Addressing bug 22090
Added <p>It is okay to use aria-valuetext on these, with or without an explicit role.</p>

Removed <code>role=math</code>

Did the rest of the recommendations, except the last two > Element is focusable, > Element that is a candidate for constraint validation but… and > iframe [seemless]
I will look at those next time.


Added role=progressbar to the meter element

I'm not sure if I got the right place for the disabled content note that Craig is talking about. I'll confirm later... in the meantime, I added it with a big ADD: and REMOVE: not on each change.