Bug 22089
Addressed issues brought up in the bug. Accepted all recommendations but did not address this yet, because I would like other opinions:

"There should probably also be a note in the document that indicates how one should be able to include static non-interactive content in a dialog by using the document role to counteract the application mode. To my knowledge, static dialog content can only be accessed in Safari with VoiceOver. I believe the Windows screen readers disallow access to any static content in dialogs, whether or not the application mode is used. I consider this to be a bug, though not everyone agrees."